refactor(worker): move the claim walk out of the routes file (#1253)
ClaimedJob, BlockedCandidate, ClaimEvaluator, collectClaimCandidates and the
evaluate-and-claim walk are now in Services/WorkerClaimEvaluation.swift.

The issue described these as nested in the routes struct. They were not - they
were file-private top-level types in WorkerJobRoutes.swift, which is the same
problem stated differently: the routes file was the only thing that could see
them, and none of them is about HTTP.

The one genuine coupling was the claim step, which lives on the route
collection because it needs the request. Rather than drag the walk back into
the routes file to reach it, it is injected as a closure.

That seam is what makes the move worth more than churn. atomicallyClaimSubmission's
own doc comment says the lost-race path "can't be triggered deterministically
through the HTTP endpoint" - it now can, and ClaimWalkTests covers three
branches that no test could previously reach:

  - a candidate another runner claimed between our scan and our claim does not
    end the walk; the next candidate is still tried
  - every candidate lost yields no job rather than a job that was not claimed
  - the claimed job carries its OWN setup and manifest, not the first
    candidate scanned - a mismatch there would hand a runner one submission
    graded against another submission's test setup

WorkerJobRoutes.swift: 701 -> 477 lines.

Verified: 542 tests across the claim, worker and runner suites, including
ClaimCompareAndSetTests which pins the atomic claim itself on both backends.
SwiftLint --strict at 0 violations in 973 files.

Item 4 of #1253.
